# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bloom",
    "version": "0.11.0"
  },
  "paths": {
...
    "/api/v1/beans": {
      "post": {
        "tags": [
          "beans"
        ],
        "summary": "Create Bean",
        "description": "Create a bean. It is shared with everyone; you are recorded as its owner.\n\n409 if this roaster already has a bean with that name (matched case-insensitively) \u2014\nopen that one, or resend with `?allow_duplicate=true` if it really is another coffee.",
        "operationId": "beans-create_bean",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "allow_duplicate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Accept a name that already exists under this roaster instead of failing with 409.",
              "default": false,
              "title": "Allow Duplicate"
            },
            "description": "Accept a name that already exists under this roaster instead of failing with 409."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeanCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeanRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "beans"
        ],
        "summary": "List Beans",
        "description": "List all beans (shared). Use `?mine=true` to return only the beans you own.",
        "operationId": "beans-list_beans",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mine",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Mine"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BeanRead"
                  },
                  "title": "Response Beans-List Beans"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/beans/{bean_id}": {
      "get": {
        "tags": [
          "beans"
        ],
        "summary": "Get Bean",
        "description": "Get a bean by id. Any authenticated user may read any bean.",
        "operationId": "beans-get_bean",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bean_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Bean Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeanRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "beans"
        ],
        "summary": "Update Bean",
        "description": "Update a bean. Only its owner (or an admin) may edit it.\n\n409 if the new name and roaster are already taken by another bean; merge into it\ninstead, or resend with `?allow_duplicate=true`.",
        "operationId": "beans-update_bean",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bean_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Bean Id"
            }
          },
          {
            "name": "allow_duplicate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Accept a name that already exists under this roaster instead of failing with 409.",
              "default": false,
              "title": "Allow Duplicate"
            },
            "description": "Accept a name that already exists under this roaster instead of failing with 409."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeanUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeanRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "beans"
        ],
        "summary": "Delete Bean",
        "description": "Delete a bean (cascades to its brews and tastings). Owner or admin only.",
        "operationId": "beans-delete_bean",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bean_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Bean Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/beans/{bean_id}/merge": {
      "post": {
        "tags": [
          "beans"
        ],
        "summary": "Merge Bean",
        "description": "Fold a duplicate into this bean: its brews, lots and recipes move here.\n\nYou must own both beans (or be an admin). This bean keeps its own values and adopts\nthe duplicate's for anything it left empty.",
        "operationId": "beans-merge_bean",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bean_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Bean Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeanMerge"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeanRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/beans/{bean_id}/recipes": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "Create Recipe",
        "description": "Create a recipe for any shared bean; you are recorded as its creator.",
        "operationId": "recipes-create_recipe",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bean_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Bean Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "recipes"
        ],
        "summary": "List Recipes",
        "description": "List a bean's recipes. Any authenticated user may read them.",
        "operationId": "recipes-list_recipes",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bean_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Bean Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipeRead"
                  },
                  "title": "Response Recipes-List Recipes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recipes/{recipe_id}": {
      "get": {
        "tags": [
          "recipes"
        ],
        "summary": "Get Recipe",
        "description": "Get a recipe by id. Any authenticated user may read it.",
        "operationId": "recipes-get_recipe",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recipe Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "recipes"
        ],
        "summary": "Update Recipe",
        "description": "Update a recipe. Only its creator (or an admin) may edit it.",
        "operationId": "recipes-update_recipe",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recipe Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "recipes"
        ],
        "summary": "Delete Recipe",
        "description": "Delete a recipe. Creator or admin only; existing brews remain unchanged.",
        "operationId": "recipes-delete_recipe",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recipe Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recipes/{recipe_id}/brews": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "Create Brew From Recipe",
        "description": "Create a brew from a recipe snapshot; any authenticated user may use it.",
        "operationId": "recipes-create_brew_from_recipe",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recipe Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrewFromRecipeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrewRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BeanCreate": {
        "properties": {
          "origin_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Country",
            "description": "Country of origin.",
            "examples": [
              "Ethiopia"
            ]
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region",
            "description": "Growing region.",
            "examples": [
              "Guji"
            ]
          },
          "producer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Producer",
            "description": "Farm or producer.",
            "examples": [
              "Tabe Burka"
            ]
          },
          "variety": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variety",
            "description": "Coffee variety.",
            "examples": [
              "Heirloom"
            ]
          },
          "process": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "washed",
                  "natural",
                  "honey",
                  "anaerobic",
                  "carbonic_maceration",
                  "other"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Process",
            "description": "Post-harvest process.",
            "examples": [
              "washed"
            ]
          },
          "roast_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "light",
                  "medium_light",
                  "medium",
                  "medium_dark",
                  "dark"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Roast Level",
            "description": "Roast level.",
            "examples": [
              "medium_light"
            ]
          },
          "roast_type": {
            "type": "string",
            "enum": [
              "filter",
              "espresso",
              "omni",
              "unknown"
            ],
            "title": "Roast Type",
            "description": "Brewing intent this coffee is roasted for.",
            "default": "unknown",
            "examples": [
              "filter"
            ]
          },
          "blend": {
            "type": "string",
            "enum": [
              "single_origin",
              "blend",
              "unknown"
            ],
            "title": "Blend",
            "description": "Single origin or a blend.",
            "default": "single_origin",
            "examples": [
              "single_origin"
            ]
          },
          "altitude_masl": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altitude Masl",
            "description": "Growing altitude (metres above sea level).",
            "examples": [
              2100
            ]
          },
          "tasting_notes_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tasting Notes Label",
            "description": "Tasting notes printed on the bag.",
            "examples": [
              "Peach, jasmine, black tea"
            ]
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating",
            "description": "Overall rating of the coffee itself, 1\u20135 (null = unrated).",
            "examples": [
              4
            ]
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website",
            "description": "URL with more info about the coffee or roaster.",
            "examples": [
              "https://nomadcoffee.es"
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Your own notes.",
            "examples": [
              "Great as filter"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Coffee name.",
            "examples": [
              "Guji Natural"
            ]
          },
          "roaster": {
            "type": "string",
            "title": "Roaster",
            "description": "Roaster name. Matched case-insensitively; created if it does not exist yet.",
            "examples": [
              "Nomad Coffee"
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "roaster"
        ],
        "title": "BeanCreate"
      },
...
      "BeanRead": {
        "properties": {
          "origin_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Country",
            "description": "Country of origin.",
            "examples": [
              "Ethiopia"
            ]
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region",
            "description": "Growing region.",
            "examples": [
              "Guji"
            ]
          },
          "producer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Producer",
            "description": "Farm or producer.",
            "examples": [
              "Tabe Burka"
            ]
          },
          "variety": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variety",
            "description": "Coffee variety.",
            "examples": [
              "Heirloom"
            ]
          },
          "process": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "washed",
                  "natural",
                  "honey",
                  "anaerobic",
                  "carbonic_maceration",
                  "other"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Process",
            "description": "Post-harvest process.",
            "examples": [
              "washed"
            ]
          },
          "roast_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "light",
                  "medium_light",
                  "medium",
                  "medium_dark",
                  "dark"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Roast Level",
            "description": "Roast level.",
            "examples": [
              "medium_light"
            ]
          },
          "roast_type": {
            "type": "string",
            "enum": [
              "filter",
              "espresso",
              "omni",
              "unknown"
            ],
            "title": "Roast Type",
            "description": "Brewing intent this coffee is roasted for.",
            "default": "unknown",
            "examples": [
              "filter"
            ]
          },
          "blend": {
            "type": "string",
            "enum": [
              "single_origin",
              "blend",
              "unknown"
            ],
            "title": "Blend",
            "description": "Single origin or a blend.",
            "default": "single_origin",
            "examples": [
              "single_origin"
            ]
          },
          "altitude_masl": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altitude Masl",
            "description": "Growing altitude (metres above sea level).",
            "examples": [
              2100
            ]
          },
          "tasting_notes_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tasting Notes Label",
            "description": "Tasting notes printed on the bag.",
            "examples": [
              "Peach, jasmine, black tea"
            ]
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating",
            "description": "Overall rating of the coffee itself, 1\u20135 (null = unrated).",
            "examples": [
              4
            ]
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website",
            "description": "URL with more info about the coffee or roaster.",
            "examples": [
              "https://nomadcoffee.es"
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Your own notes.",
            "examples": [
              "Great as filter"
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "examples": [
              1
            ]
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "Owner id (who added the bean).",
            "examples": [
              1
            ]
          },
          "owner": {
            "$ref": "#/components/schemas/AuthorRead",
            "description": "Owner (who added the bean)."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Guji Natural"
            ]
          },
          "roaster": {
            "$ref": "#/components/schemas/RoasterRead",
            "description": "The roaster this bean came from."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "examples": [
              "2026-07-05T09:30:00Z"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "owner",
          "name",
          "roaster",
          "created_at"
        ],
        "title": "BeanRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        }
      }
    }
  }
}